% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.distance.r
\name{edge.distance}
\alias{edge.distance}
\title{Edge Distance}
\usage{
edge.distance(g, edge.network = NULL, add.names = T)
}
\arguments{
\item{g}{An igraph graph}

\item{edge.network}{A pre-calculated edge network from ```get.edge.network```}
}
\description{
Calculates the distance from each node in the network to each edge in the network.
}
\examples{
get.edge.network(g)
}

