% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edge.distance2}
\alias{edge.distance2}
\title{Edge Distance}
\usage{
edge.distance2(g, edge.network = NULL, add.names = T)
}
\arguments{
\item{g}{An igraph graph}
}
\description{
Calculates the distance from each node in the network to each edge in the network.
}
\examples{
get.edge.network(g)
}

