% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii.r
\docType{data}
\name{g.samp}
\alias{g.samp}
\title{The Sampson Monastery network.}
\format{An igraph object with 18 nodes and 90 edges:
\describe{
  \item{name}{name of the monk}
  \item{valence}{valence of the edge}
}}
\source{
\url{https://sites.google.com/site/ucinetsoftware/datasets/sampsonmonastery}
}
\usage{
g.samp
}
\description{
This network contains nodes representing monks in a monastery.
}
\keyword{datasets}

