% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pii}
\alias{pii}
\title{Political Independence Index}
\usage{
pii(g, pii.beta = -0.8, e.dist = NULL)
}
\arguments{
\item{g}{An igraph graph}

\item{pii.beta}{Should the vertex and edge names be added to the rows and columns of the matrix.}

\item{e.dist}{(optional) an edge.distance matrix, if calculated ahead of time.}
}
\description{
Returns a numeric vector of PII scores for each vertex in the network.
}
\examples{
pii(g)
}

