% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pii2}
\alias{pii2}
\title{Political Independence Index}
\usage{
pii2(g, pii.beta = -0.8, e.dist = NULL, triadic = F, pii.delta = 1,
  e.triads = NULL)
}
\arguments{
\item{g}{An igraph graph}

\item{pii.beta}{Should the vertex and edge names be added to the rows and columns of the matrix.}

\item{e.dist}{(optional) an edge.distance matrix, if calculated ahead of time.}
}
\description{
Returns a numeric vector of PII scores for each vertex in network.
}
\examples{
pii2(g)
}

