% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.table.r
\name{triad.table}
\alias{triad.table}
\title{Triad Table}
\usage{
triad.table(g)
}
\arguments{
\item{g}{An igraph graph}
}
\description{
Creates a table that contains triadID - the three nodes that make up the triad, nodeID - the character name of the node,
direction of the triad from the node, the valence of the triad from the node's perspective, distance of the triad from the node,
and nodeNum - the numderic value of the node
}

