% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triadic.edges.r
\name{triadic.edges}
\alias{triadic.edges}
\title{Triadic Edges}
\usage{
triadic.edges(g)
}
\arguments{
\item{g}{An igraph graph}
}
\description{
Identifies which edges in a network are part of a triad.
}
\examples{
get.edge.network(g)
}
