% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vid.r
\name{vid}
\alias{vid}
\title{Vertex ID}
\usage{
vid(g, x)
}
\arguments{
\item{g}{An igraph graph}

\item{x}{A vertex name (character)}
}
\description{
Returns the vertex ID when given a name and a graph.
}
\examples{
vid(g, x="Billy")
}

