% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edge.distance}
\alias{edge.distance}
\title{Edge Distance}
\usage{
edge.distance(g, add.names = T)
}
\arguments{
\item{g}{An igraph graph}

\item{add.names}{Should the vertex and edge names be added to the rows and columns of the matrix.}
}
\description{
Returns a rectangular matrix where the rows are the edges of the graph, and the columns are the vertices of the graph.
The cells of the matrix is the distance from the vertex to the edge. The edges directly incident on the vertex are distance 0.
It is used in the calculation of PII.
}
\examples{
edge.distance(g)
}

