% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.edge.network}
\alias{get.edge.network}
\title{Get Edge Network}
\usage{
get.edge.network(g)
}
\arguments{
\item{g}{An igraph graph}
}
\description{
Returns a modfied network where each edge in the original network is split in two and a new node is inserted in between and given the label of the original edge.
}
\examples{
get.edge.network(g)
}

