% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piiCalc}
\alias{piiCalc}
\title{Calculate the PII scores after the edge distances are calculated}
\usage{
piiCalc(edgeDistance, edgevalence, piiBeta, piiX, maxDistance)
}
\arguments{
\item{edgeDistance}{An integer matrix of distances from the nodes to the edges}

\item{piiBeta}{The beta attenuation value (ideally -1.0 to -0.01)}

\item{piiX}{The normalization factor based on the maximum degree.}

\item{maxDistance}{The maximum distance in the edgeDistance matrix (or some user-provided max distance for very large networks)}

\item{valence}{The postive / negative valence of each of the edges.}
}
\description{
Calculate the PII scores after the edge distances are calculated
}

